Project's now have access to their configs
