+++ lib/xcode/target.rb
@@ -52,70 +52,6 @@ module Xcode
     # @return [Project] the reference to the project for which these targets reside.
     attr_accessor :project
     
-    #
-    # @return [Array<BuildConfiguration>] the configurations that this target 
-    #   supports. These are generally 'Debug' or 'Release' but may be custom 
-    #   created configurations.
-    # 
-    def configs
-      build_configuration_list.build_configurations.map do |config|
-        config.target = self
-        config
-      end
-    end
-    
-    #
-    # Return a specific build configuration. 
-    # 
-    # @note an exception is raised if no configuration matches the specified name.
-    # 
-    # 
-    # @param [String] name of a configuration to return
-    # 
-    # @return [BuildConfiguration] a specific build configuration that 
-    #   matches the specified name.
-    #
-    def config(name)
-      config = configs.select {|config| config.name == name.to_s }.first
-      raise "No such config #{name}, available configs are #{configs.map {|c| c.name}.join(', ')}" if config.nil?
-      yield config if block_given?
-      config
-    end
-    
-    #
-    # Create a configuration for the target.
-    # 
-    # @example debug configuration
-    # 
-    #   target.create_config 'Debug' do |config|
-    #     # configuration the new debug config.
-    #   end
-    # 
-    # @param [String] name of the configuration to create
-    # @return [BuildConfiguration] that is created
-    #
-    def create_configuration(name)
-      # To create a configuration, we need to create or retrieve the configuration list
-      
-      created_config = build_configuration_list.create_config(name) do |config|
-        yield config if block_given?
-      end
-      
-      created_config
-    end
-
-    def create_configurations(*configuration_names)
-      
-      configuration_names.compact.flatten.map do |config_name|
-        created_config = create_configuration config_name do |config|
-          yield config if block_given?
-        end
-        
-        created_config.save!
-      end
-      
-    end
-    
     # 
     # @return [BuildPhase] the framework specific build phase of the target.
     #
